+++ src/ServicesProvider/ServicesProvider.php
@@ -1,40 +1,40 @@
 <?php
 
 /*
- * UF Breadcrumb
+ * UserFrosting Breadcrumb Sprinkle
  *
- * @link https://github.com/lcharette/UF_Breadcrumb
- *
- * @copyright Copyright (c) 2016 Louis Charette
- * @license https://github.com/lcharette/UF_Breadcrumb/blob/master/LICENSE (MIT License)
+ * @link      https://github.com/lcharette/UF_Breadcrumb
+ * @copyright Copyright (c) 2020 Louis Charette
+ * @license   https://github.com/lcharette/UF_Breadcrumb/blob/master/LICENSE (MIT License)
  */
 
 namespace UserFrosting\Sprinkle\Breadcrumb\ServicesProvider;
 
-use UserFrosting\Sprinkle\Breadcrumb\BreadcrumbManager;
+use Psr\Container\ContainerInterface;
+use UserFrosting\Sprinkle\Breadcrumb\Breadcrumb\Manager;
 use UserFrosting\Sprinkle\Breadcrumb\Twig\BreadcrumbExtension;
 
 /**
  * Registers services for the Breadcrumb sprinkle.
- *
- * @author Louis Charette (https://github.com/lcharette)
  */
 class ServicesProvider
 {
     /**
      * Register UserFrosting's Breadcrumb services.
      *
-     * @param Container $container A DI container implementing ArrayAccess and container-interop.
+     * @param ContainerInterface $container A DI container implementing ArrayAccess and container-interop.
      */
-    public function register($container)
+    public function register(ContainerInterface $container)
     {
         /*
          * Authentication service.
          *
          * Supports logging in users, remembering their sessions, etc.
+         *
+         * @return \UserFrosting\Sprinkle\Breadcrumb\Breadcrumb\Manager
          */
         $container['breadcrumb'] = function ($c) {
-            return new BreadcrumbManager($c);
+            return new Manager($c->config, $c->translator, $c->router);
         };
 
         /*
@@ -44,7 +44,7 @@ class ServicesProvider
          */
         $container->extend('view', function ($view, $c) {
             $twig = $view->getEnvironment();
-            $extension = new BreadcrumbExtension($c);
+            $extension = new BreadcrumbExtension($c->breadcrumb);
             $twig->addExtension($extension);
 
             return $view;