+++ src/Twig/BreadcrumbExtension.php
@@ -1,41 +1,49 @@
 <?php
 
 /*
- * UF Breadcrumb
+ * UserFrosting Breadcrumb Sprinkle
  *
- * @link https://github.com/lcharette/UF_Breadcrumb
- *
- * @copyright Copyright (c) 2016 Louis Charette
- * @license https://github.com/lcharette/UF_Breadcrumb/blob/master/LICENSE (MIT License)
+ * @link      https://github.com/lcharette/UF_Breadcrumb
+ * @copyright Copyright (c) 2020 Louis Charette
+ * @license   https://github.com/lcharette/UF_Breadcrumb/blob/master/LICENSE (MIT License)
  */
 
 namespace UserFrosting\Sprinkle\Breadcrumb\Twig;
 
-use Interop\Container\ContainerInterface;
+use Twig\Extension\AbstractExtension;
+use Twig\Extension\GlobalsInterface;
+use UserFrosting\Sprinkle\Breadcrumb\Breadcrumb\Manager;
 
 /**
  * Extends Twig functionality for the Breadcrumb sprinkle.
- *
- * @author Louis Charette (https://github.com/lcharette)
  */
-class BreadcrumbExtension extends \Twig_Extension
+class BreadcrumbExtension extends AbstractExtension implements GlobalsInterface
 {
-    protected $services;
+    /**
+     * @var Manager
+     */
+    protected $breadcrumb;
 
-    public function __construct(ContainerInterface $services)
+    public function __construct(Manager $breadcrumb)
     {
-        $this->services = $services;
+        $this->breadcrumb = $breadcrumb;
     }
 
+    /**
+     * @return string
+     */
     public function getName()
     {
         return 'userfrosting/breadcrumb';
     }
 
+    /**
+     * @return array<string,mixed>
+     */
     public function getGlobals()
     {
         return [
-            'breadcrumbs'   => $this->services['breadcrumb']->get(),
+            'breadcrumbs' => $this->breadcrumb->generate(),
         ];
     }
 }